+++ src/main/java/com/atlauncher/gui/card/PackCard.java
@@ -37,7 +37,6 @@ import javax.swing.JTextArea;
 import java.awt.BorderLayout;
 import java.awt.Dimension;
 import java.awt.FlowLayout;
-import java.awt.GridBagConstraints;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
@@ -54,6 +53,7 @@ public class PackCard extends CollapsiblePanel implements RelocalizationListener
     private final JButton supportButton = new JButton(Language.INSTANCE.localize("common.support"));
     private final JButton websiteButton = new JButton(Language.INSTANCE.localize("common.website"));
     private final JButton modsButton = new JButton(Language.INSTANCE.localize("pack.viewmods"));
+    private final JButton removePackButton = new JButton(Language.INSTANCE.localize("pack.removepack"));
     private final JPanel actionsPanel = new JPanel(new BorderLayout());
     private final JSplitPane splitter = new JSplitPane();
     private final Pack pack;
@@ -67,12 +67,18 @@ public class PackCard extends CollapsiblePanel implements RelocalizationListener
         this.splitter.setRightComponent(this.actionsPanel);
         this.splitter.setEnabled(false);
 
-        JPanel abPanel = new JPanel(new FlowLayout(FlowLayout.CENTER));
-        abPanel.add(this.newInstanceButton);
-        abPanel.add(this.createServerButton);
-        abPanel.add(this.supportButton);
-        abPanel.add(this.websiteButton);
-        abPanel.add(this.modsButton);
+        JPanel top = new JPanel(new FlowLayout());
+        JPanel bottom = new JPanel(new FlowLayout());
+        JSplitPane as = new JSplitPane(JSplitPane.VERTICAL_SPLIT);
+        as.setEnabled(false);
+        as.setTopComponent(top);
+        as.setBottomComponent(bottom);
+        top.add(this.newInstanceButton);
+        top.add(this.createServerButton);
+        bottom.add(this.supportButton);
+        bottom.add(this.websiteButton);
+        bottom.add(this.modsButton);
+        bottom.add(this.removePackButton);
 
         this.descArea.setText(pack.getDescription());
         this.descArea.setLineWrap(true);
@@ -82,7 +88,7 @@ public class PackCard extends CollapsiblePanel implements RelocalizationListener
 
         this.actionsPanel.add(new JScrollPane(this.descArea, JScrollPane.VERTICAL_SCROLLBAR_AS_NEEDED, JScrollPane
                 .HORIZONTAL_SCROLLBAR_NEVER), BorderLayout.CENTER);
-        this.actionsPanel.add(abPanel, BorderLayout.SOUTH);
+        this.actionsPanel.add(as, BorderLayout.SOUTH);
         this.actionsPanel.setPreferredSize(new Dimension(this.actionsPanel.getPreferredSize().width, 180));
 
         this.getContentPane().add(this.splitter);
@@ -95,6 +101,10 @@ public class PackCard extends CollapsiblePanel implements RelocalizationListener
 
         if (!this.pack.canCreateServer()) {
             this.createServerButton.setVisible(false);
+        }
+
+        if (!this.pack.isSemiPublic() || this.pack.isTester()) {
+            this.removePackButton.setVisible(false);
         }
     }
 
@@ -109,13 +119,13 @@ public class PackCard extends CollapsiblePanel implements RelocalizationListener
                 if (App.settings.isInOfflineMode()) {
                     String[] options = {Language.INSTANCE.localize("common.ok")};
                     JOptionPane.showOptionDialog(App.settings.getParent(), Language.INSTANCE.localize("pack" + "" +
-                                    ".offlinenewinstance"), Language.INSTANCE.localize("common.offline"), JOptionPane
+                            ".offlinenewinstance"), Language.INSTANCE.localize("common.offline"), JOptionPane
                             .DEFAULT_OPTION, JOptionPane.ERROR_MESSAGE, null, options, options[0]);
                 } else {
                     if (App.settings.getAccount() == null) {
                         String[] options = {Language.INSTANCE.localize("common.ok")};
                         JOptionPane.showOptionDialog(App.settings.getParent(), Language.INSTANCE.localize("instance"
-                                + ".cannotcreate"), Language.INSTANCE.localize("instance.noaccountselected"),
+                                        + ".cannotcreate"), Language.INSTANCE.localize("instance.noaccountselected"),
                                 JOptionPane.DEFAULT_OPTION, JOptionPane.ERROR_MESSAGE, null, options, options[0]);
                     } else {
                         new InstanceInstallerDialog(pack);
@@ -135,7 +145,7 @@ public class PackCard extends CollapsiblePanel implements RelocalizationListener
                     if (App.settings.getAccount() == null) {
                         String[] options = {Language.INSTANCE.localize("common.ok")};
                         JOptionPane.showOptionDialog(App.settings.getParent(), Language.INSTANCE.localize("instance"
-                                + ".cannotcreate"), Language.INSTANCE.localize("instance.noaccountselected"),
+                                        + ".cannotcreate"), Language.INSTANCE.localize("instance.noaccountselected"),
                                 JOptionPane.DEFAULT_OPTION, JOptionPane.ERROR_MESSAGE, null, options, options[0]);
                     } else {
                         new InstanceInstallerDialog(pack, true);
@@ -162,6 +172,13 @@ public class PackCard extends CollapsiblePanel implements RelocalizationListener
                 new ViewModsDialog(pack).setVisible(true);
             }
         });
+
+        this.removePackButton.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                App.settings.removePack(pack.getCode());
+            }
+        });
     }
 
     @Override
@@ -171,5 +188,6 @@ public class PackCard extends CollapsiblePanel implements RelocalizationListener
         this.supportButton.setText(Language.INSTANCE.localize("common.support"));
         this.websiteButton.setText(Language.INSTANCE.localize("common.website"));
         this.modsButton.setText(Language.INSTANCE.localize("pack.viewmods"));
+        this.removePackButton.setText(Language.INSTANCE.localize("pack.removepack"));
     }
 }