+++ CHANGELOG.md
@@ -10,3 +10,4 @@ Changelog
 - Fix issues with old Java paths causing issues launching packs
 - Some changes to logging to the console to keep things in order under large loads
 - Add a way for semi public packs to auto add their codes to users launcher
+- Add the remove pack button back in for semi public packs